backend/requirements.txt
uvicorn[standard]
fastapi
SQLAlchemy[asyncio]
pydantic-settings
asyncpg
psycopg2-binary
pgvector
langchain-community
sentence-transformers
boto3
google-generativeai
python-multipart
